+++ application/views/TitleHistory.php
@@ -4,7 +4,7 @@
 </div>
 <?php } ?>
 
-<table class="table table-striped table-bordered">
+<table class="table table-striped table-bordered tablesorter">
 	<thead>
 		<tr>
 			<th class="col-md-2">Updated at</th>